Phase 1 complete: Extract cosmic events and game config data
