+++ memory/tasks/TASK012-implement-tech-debt-refactors.md
@@ -50,17 +50,17 @@ Following the spec-driven workflow with 5 phases aligned to DESIGN002:
 
 ## Progress Tracking
 
-**Overall Status:** In Progress - 5%
+**Overall Status:** In Progress - 25%
 
 ### Phase 1: Data and Configuration Extraction
 
 | ID | Description | Status | Updated | Notes |
 | --- | --- | --- | --- | --- |
-| 1.1 | Extract cosmic events to /src/data/cosmicEvents.ts | Not Started | 2025-10-26 | Priority 1 |
-| 1.2 | Create cosmicEventResolver in /src/lib/gameplay/ | Not Started | 2025-10-26 | Priority 1 |
-| 1.3 | Move UPGRADE_CONFIG to /src/data/gameConfig.ts | Not Started | 2025-10-26 | Priority 1 |
-| 1.4 | Move FRACTAL_ZONES to /src/data/gameConfig.ts | Not Started | 2025-10-26 | Priority 1 |
-| 1.5 | Validate Phase 1 with tests | Not Started | 2025-10-26 | Checkpoint |
+| 1.1 | Extract cosmic events to /src/data/cosmicEvents.ts | Complete | 2025-10-26 | Successfully extracted |
+| 1.2 | Create cosmicEventResolver in /src/lib/gameplay/ | Complete | 2025-10-26 | Utility functions exported |
+| 1.3 | Move UPGRADE_CONFIG to /src/data/gameConfig.ts | Complete | 2025-10-26 | With types |
+| 1.4 | Move FRACTAL_ZONES to /src/data/gameConfig.ts | Complete | 2025-10-26 | With types |
+| 1.5 | Validate Phase 1 with tests | Complete | 2025-10-26 | All 22 tests passing ✅ |
 
 ### Phase 2: Utility and Helper Extraction
 
@@ -137,3 +137,15 @@ Following the spec-driven workflow with 5 phases aligned to DESIGN002:
 - Created detailed subtask breakdown (40+ subtasks)
 - Set priority levels for incremental delivery
 - Ready to begin Phase 1 implementation
+
+**Phase 1 Completed:**
+- Extracted cosmic events catalog to /src/data/cosmicEvents.ts
+- Created cosmic event resolver in /src/lib/gameplay/cosmicEventResolver.ts
+- Exported clampWeight and meetsRequirement utility functions for testability
+- Moved UPGRADE_CONFIG to /src/data/gameConfig.ts with types
+- Moved FRACTAL_ZONES to /src/data/gameConfig.ts with FractalZone type
+- Moved DIMENSIONAL_TARGET to gameConfig
+- Updated cosmicEvents.ts to re-export from new modules for backward compatibility
+- Updated starthere.tsx imports to use new configuration modules
+- Validated: All 22 tests passing, linting clean (only pre-existing warnings)
+- Phase 1 complete - data extraction successful with zero breaking changes
