+++ src/app/pages/starthere.tsx
@@ -22,80 +22,7 @@ import FractalViewer, { type FractalRendererMode } from "./fractalviewer";
 import StartHereMenu from "./startheremenu";
 import { resolveCosmicEvent } from "./cosmicEvents";
 import { useGameStore, type UpgradeKey, type ComplexParameter } from "@/store/gameStore";
-
-type UpgradeConfig = {
-  title: string;
-  description: string;
-  baseCost: number;
-  growth: number;
-  flavor: string;
-};
-
-const UPGRADE_CONFIG: Record<UpgradeKey, UpgradeConfig> = {
-  probe: {
-    title: "Fractal Probes",
-    description:
-      "Deploy autonomous explorers that map detail while you plan the next zoom.",
-    baseCost: 35,
-    growth: 1.45,
-    flavor: "Their quantum lenses never blink.",
-  },
-  processor: {
-    title: "Quantum Processors",
-    description:
-      "Accelerate every calculation, multiplying the insight gathered by your probes.",
-    baseCost: 120,
-    growth: 1.6,
-    flavor: "The algorithms hum in complex harmony.",
-  },
-  stabilizer: {
-    title: "Dimensional Stabilizers",
-    description:
-      "Anchor deeper zoom levels so the structure holds while automation continues.",
-    baseCost: 220,
-    growth: 1.55,
-    flavor: "Reality threads itself tighter around the frontier.",
-  },
-};
-
-const FRACTAL_ZONES = [
-  {
-    name: "Mandelbrot Core",
-    requirement: 0,
-    bonus: 0.0,
-    description: "The familiar heart of the set, steady and welcoming.",
-  },
-  {
-    name: "Seahorse Valley",
-    requirement: 4,
-    bonus: 0.08,
-    description:
-      "Filigree spirals curl like cosmic seahorses, enriching every data packet.",
-  },
-  {
-    name: "Spiral Nebula",
-    requirement: 7,
-    bonus: 0.18,
-    description:
-      "Twin galaxies of recursion feed each other, boosting production dramatically.",
-  },
-  {
-    name: "Mini-Brot Frontier",
-    requirement: 11,
-    bonus: 0.3,
-    description:
-      "Self-similarity upon self-similarity. You glimpse new universes in each pixel.",
-  },
-  {
-    name: "Hyperbolic Bloom",
-    requirement: 15,
-    bonus: 0.45,
-    description:
-      "Geometries bloom beyond Euclid. Ascension-grade knowledge saturates every line.",
-  },
-];
-
-const DIMENSIONAL_TARGET: ComplexParameter = { real: -0.75, imaginary: 0.11 };
+import { UPGRADE_CONFIG, FRACTAL_ZONES, DIMENSIONAL_TARGET } from "@/data/gameConfig";
 
 const formatNumber = (value: number): string => {
   if (value >= 1e9) {
